Fixture: restore, add tree-sitter
